requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Optional - Form Data Handling
python-multipart==0.0.6

# CORS & Security
python-dotenv==1.0.0
